Migrate to proc_copy_credentials
proc_copy_credentials does not update the uthread credentials, but
neglecting to do so has not shown any negative effects so far.

Unfortunately, because libmemctl is not thread safe, I had to disable
the second bootstrap MITM thread. Honestly I'm not sure whether there
are supposed to be two threads or not, given that there's only one
global variable to store the thread ID. However, disabling the second
bootstrap MITM thread has not shown any adverse effects so far.